backend/pyproject.toml
[project]
name = "music-assistant-backend"
version = "1.0.0"
description = "Диалоговая система с поддержкой естественного языка. Предметная область: Музыка"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.22",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.14.0",
    "python-dotenv>=1.0.0",
    "firebase-admin>=6.5.0",
    "langchain>=0.3.0",
    "langchain-groq>=0.2.0",
    "langgraph>=0.4.0",
    "langchain-core>=0.3.0",
    "pydantic-settings>=2.5.0",
    "python-multipart>=0.0.9",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008", "E712"]

[tool.ruff.lint.isort]
known-first-party = ["dispatch"]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
]
